feature: add seperate file
