+++ src/Hooks/useSearch.ts
@@ -1,0 +1,10 @@
+import { useContext } from "react";
+import { SearchContext } from "@/Context/SearchContext";
+
+export const useSearch = () => {
+    const context = useContext(SearchContext);
+    if (!context) {
+        throw new Error("useSearch must be used within a SearchProvider");
+    }
+    return context;
+};
